move components in post, account and header folders
