+++ front/src/components/account/PasswordValidator.js
