+++ front/src/components/account/UserInfo.js
@@ -3,7 +3,7 @@ import React, { useState, useEffect, useContext } from 'react'
 import { useNavigate } from 'react-router-dom'
 
 import axios from 'axios';
-import { LocalContext } from '../Context/LocalContext';
+import { LocalContext } from '../../context/LocalContext';
