+++ front/src/index.js
@@ -3,13 +3,10 @@ import ReactDOM from 'react-dom/client';
 import './index.css';
 import App from './App';
 import { BrowserRouter } from 'react-router-dom';
-import LocalContextProvider from './Context/LocalContext';
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
 root.render(
-  <LocalContextProvider>
     <BrowserRouter>
       <App />
     </BrowserRouter>
-  </LocalContextProvider>
 );
